+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # To be released:
 - Removed many internal implementation classes and methods from the SDK's public API
 
+- Fix sending GIFs from keyboard
+
 # Oct 16th, 2020 - 4.3.1-beta-1
 - Significant performance improvements
 - Fix a crash related to behaviour changes in 1.3.0-alpha08 of the AndroidX Fragment library